Display Unicode emoji characters
This change:
* Modifies the EmojiCollection type to carry both the emoji text name
(as it did before) as well as the Unicode text to which it corresponds.
The Unicode text is obtained by parsing the "filename" field of each
emoji JSON record and interpreting it as a series of Unicode code point
values delimited by "-", e.g. "1f6c0-1f3ff".
* Modifies various parts of the UI, as well as Markdown rendering, to
resolve any emoji names (e.g. ":foo:") to their Unicode text.