Add partial support for detecting post permalinks
This change adds a new LinkTarget type, LinkPostPermalink, as well as
logic to match message URLs against the server's base URL in order to
detect post permalinks. This feature is not complete because the links
do not get opened; this patch just starts to flesh out adding support
for this feature.